pyproject.toml
[tool.poetry]
name = "haikubot"
version = "0.1.0"
description = ""
authors = ["Valeriy Pogorelov <valeriy.pogorelov@together.com>"]
readme = "README.md"

[tool.poetry.scripts]
start-bot = "haikubot:main"
init-db = "init_db:main"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
openai = "^1.57.0"
python-telegram-bot = "^21.8"
supabase = "^2.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
